递归默认配置
