+++ ID-SDK/src/main/resources/conf/config.json
@@ -1,7 +1,7 @@
 {
-   "ip":"45.120.243.40",
+   "ip":"36.112.25.8",
    "port":"3641",
    "query":true,
-   "admin":true,
+   "admin":false,
    "protocol":"TCP"
 }
