+++ ID-SDK/src/main/resources/public_key.pem
@@ -1,5 +1,8 @@
 -----BEGIN PUBLIC KEY-----
-MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDJ9Fpee/792X7cxHoiJuRj/iha/Ab9ldCr48Tm
-9pZJdy/Jucc+kft+9QqQZaMSzMnstoShFsCWSKGMeu2mG5uP9CwvBoI9gZJGAVqzSWY9+3lxe2Bj
-Tszjq6nwWzCS5ayIjVUpAnnFqxWS2W44fl9p1VtEZuPCCllt7Fd2ZZAFTwIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAnlek5Q99/6QQSsfezTSDhj1jlc66KnCK
+B7LooKvrglfXijBbhU0wunhN44rGm/NNy344eZp63j8fo85YZAmfr8JxmBxqP7PGv78/AdBZay3U
+drRV3Nxd++AFAyLI5JPVh7ssNovDzfKbX/cR+Qpku2MJ4XuES/UaathwEzCoiwbxdmHC2fSsK2oO
+9LbwNa8+1CnMSgP43f3zgt7j4rbk/oEqrGlYKvMsZ7UkSlF6SLLWxEBx4EEB4/V8orM3CJXrQySB
+Vxk1agQX1e909Q7ve5IUq5Ecje5s7kRVwE8kKJamZkxz1M3SMvackcE/Q46cRhOVHFrMq0MVA19u
+13ZgJQIDAQAB
 -----END PUBLIC KEY-----
